refactor(ui): extract ViewSwitch types and styles into dedicated files, use theme spacing tokens
