+++ redisinsight/ui/src/components/browser/view-switch/index.ts
@@ -1,2 +1,2 @@
 export { default as ViewSwitch } from './ViewSwitch'
-export type { ViewSwitchProps } from './ViewSwitch'
+export type { ViewSwitchProps } from './ViewSwitch.types'
